Deleted Redundant Decision
For some reason the duplication of the call to a vanilla event modifier makes the decisions listbox scroll go bork so I deleted the duplicated decision and since it gives the same bonuses to the tag as the vanilla one no a problem deleting it. Probably the person who made this didn't pay attention when moving the decision to this file.